addressing matbun and okrochak PR comments
