+++ src/itwinai/model_hub/manifest.py
